animation
created a base class for the shared behaviours of character agents in the scene which i intend to bring the player character under. currently stores animations in a map, speed, direction and a moveTowards method that I want to rig up to cover both npc movement and player movement. using this class i won't have to create like 10 fields to store each animation, and the dev won't need to create a new class just to give something a different set of animations.