states and turning
the npc exists in a given state which determines its behaviour. currently patrolling, searching, and pursuing the player, with state transitions between.
the npc now takes time to turn around, which telegraphs the vision cone to the player instead of it just snapping onto them.
left some comment shrapnel because i originally used another method which was cool but it meant the npc would stop to turn and face you whilst chasing you, and also felt kind of robotic because it would just stand there and swivel around between movements, so now the npc can turn and move simultaneously. this function can also be used to have the npc just look around, and it not only makes the game more fair on the player but makes the npc seem more natural and human.
earlier today i made a change so that the npc could see over low cover, but i removed it since it was harder for the player to tell when they are and are not visible to the npc, so i ditched it for clarity.