movement changes and refactoring
running animations keep their playback position when changing directions, player uses acceleration and deceleration, reworking player movement code, got picking up items working on this branch also. code for npcs decelerating to their destination is in there but commented because they got a little slow on corners. using a charactermovementcomponent for both player and npcs which allowed me to add the field to superclass. removed a bunch of unnecessary fields so they could be inherited instead