updated HUD from main
yeah idk why, but somehow after I merged in callums cool changes to ui, I apparently edited the old ui to work with the new npc stuff, and so it just stayed as old ui. like idk why the merge didn't change what UI I was working on, maybe i didn't close unreal or something and it didn't update in realtime, but yeah now I've brought it in. not hooked up atm