both customer flipbooks made, customer AI and social stealth
automated customer spawning, randomly chooses male or female sprites on spawn, bottom left UI changed because all it did previously was indicate the same thing as the bottom right except worse, now it tells the player if they are being seen. lines are drawn pointing to observers, coloured to convey suspicion since it's easy to miss the battery bar changing, plus there is nothing indicating what his battery UI even represents. npcs won't get stuck on eachother, when concealing an item is completed, observers will have a chance of setting the suspicion to 100 based on how few other customers are in their LOS. turned off npc footsteps since it gets ridiculous when there are like 15 customers in the scene. fixed the spawner too, moveable and no longer crashes. detect will return true if the subject is really close to the npc, within 100 units. npcs now stare at a concealing player, so if you're seen in the act of concealing you will either commit crime in their vision or cancel it in time. added more clerks but level still feels too easy to avoid npcs because of how big it is, and now we have to make more assets because of this decision too? like he's just inventing his own problems now and getting mald at us for not solving them. customers spawn every 5 seconds until limit of 30 is reached, not implemented customer exiting as of yet, and the one-way gates are only going to cause issues with npcs trying to path through the wrong side of it and getting stuck